+++ src/lib/pages/proposal-details/components/proposal-overview/proposal-period-overview/index.tsx
@@ -4,6 +4,7 @@ import type { ProposalOverviewProps } from "..";
 import { ProposalStepper } from "../../proposal-stepper";
 
 import { DepositOverview } from "./DepositOverview";
+import { VotingOverview } from "./VotingOverview";
 
 export const ProposalPeriodOverview = ({
   proposalData,
@@ -16,5 +17,6 @@ export const ProposalPeriodOverview = ({
     <ProposalStepper step={1} proposalData={proposalData} isOverview />
     <DepositOverview proposalData={proposalData} {...props} />
     <ProposalStepper step={2} proposalData={proposalData} isOverview />
+    <VotingOverview proposalData={proposalData} {...props} />
   </div>
 );