+++ CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#778](https://github.com/alleslabs/celatone-frontend/pull/778) Proposal period overview voting
 - [#775](https://github.com/alleslabs/celatone-frontend/pull/775) Proposal period overview deposit
 - [#766](https://github.com/alleslabs/celatone-frontend/pull/766) Proposal status summary body
 - [#763](https://github.com/alleslabs/celatone-frontend/pull/763) Proposal status summary top